requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
requests==2.32.4
numpy==2.3.2
pandas==2.3.1
pymupdf==1.26.3
faiss-cpu==1.11.0.post1
rank-bm25==0.2.2
sentence-transformers==5.0.0
scikit-learn==1.7.1
orjson==3.11.1
openpyxl==3.1.5